+++ changelog.md
@@ -1,5 +1,11 @@
 # changelog
 
+## 23.0.3
+
+- fixed reading
+
+- "read_str" and "read_file" renamed to "from_str" and "from_file"
+
 ## 23.0.2
 
 - added ability to pass expected type
@@ -11,7 +17,7 @@
 ## 23.0.0
 
 - fixed typing
-- changed arg name from "file_patg" to "path"
+- "file_path" renamed to "path"
 
 ## 22.0.5